refactor: move module for consistency to config
