+++ src/libs/satellite/src/auth/alternative_origins.rs
@@ -1,4 +1,4 @@
-use crate::auth::types::state::AuthenticationConfig;
+use crate::auth::types::config::AuthenticationConfig;
 use crate::storage::store::get_custom_domains_store;
 use crate::storage::strategy_impls::StorageState;
 use ic_cdk::id;
